+++ release.nix
@@ -1,3 +1,8 @@
+# © 2019 Serokell <hi@serokell.io>
+# © Lars Jellema <lars.jellema@gmail.com>
+#
+# SPDX-License-Identifier: MPL-2.0
+
 let
   pkgs = import <nixpkgs> { };
   inherit (pkgs) haskell;